add food + delete food controller + multer

+++ backend/controllers/foodController.js
@@ -7,7 +7,7 @@ import fs from "fs"
 const listFood = async (req, res) => {
     try {
         const foods = await foodModel.find({})
-        res.json({ success: true, data: foods })
+        res.json({ success: true, foods })
     } catch (error) {
         console.log(error);
         res.json({ success: false, message: "Error" })
@@ -52,14 +52,24 @@ const addFood = async (req, res) => {
 // delete food
 const removeFood = async (req, res) => {
     try {
-
+        const {id} = req.body;
+        if(!id){
+                return res.json({ success: false, message: "Select food to delete or provide id" })
+        }
         const food = await foodModel.findById(req.body.id);
+
+        if(!food){
+                return res.json({ success: false, message: "Food already removed" })
+        }
+
+        await cloudinary.uploader.destroy(food.image.public_id);
 
         await foodModel.findByIdAndDelete(req.body.id)
         res.json({ success: true, message: "Food Removed" })
 
     } catch (error) {
-        res.json({ success: false, message: "Error" })
+        console.log(error)
+        res.json({ success: false, message: error.message })
     }
 
 }
